Rename config.py to settings.py for consistency
