+++ src/wg_discord/settings.py
